requirements.txt
aiogram==3.0.0b3
aiohttp==3.8.3
openai==0.27.0
python-dotenv==0.19.2
asyncpg==0.25.0
sqlalchemy==1.4.29
aiopg